// src/app/main/Modules/produit/Model/Product.ts
export interface products 
    {
        idProduit: number
        libelleProduit: string
        prix: number
        dateCreation:Date
        dateDernierModification:Date
        archive: boolean
        categorieProduit: string
        detailFactures: string
        codeProduit: string
        stock: number
        image:string
    
    }